Fixed enum stuff
